реализация зависимостей
