+++ путь_к_файлу.txt
@@ -1,0 +1,10 @@
+третье слово
+третье слово
+третье слово
+третье слово
+третье слово
+третье слово
+третье слово
+третье слово
+третье слово
+третье слово
